+++ core/src/test/java/de/jetsli/graph/trees/QuadTreeSimple8Test.java
@@ -22,7 +22,7 @@ package de.jetsli.graph.trees;
 public class QuadTreeSimple8Test extends QuadTreeTester {
 
     @Override
-    protected QuadTree createQuadTree(int items) {
+    protected QuadTree createQuadTree(long items) {
         return new QuadTreeSimple(8).init(items);
     }
 }
