+++ core/src/test/java/de/jetsli/graph/trees/QuadTreeSimpleTest.java
@@ -25,7 +25,7 @@ import static org.junit.Assert.*;
 public class QuadTreeSimpleTest extends QuadTreeTester {
 
     @Override
-    protected QuadTree createQuadTree(int items) {
+    protected QuadTree createQuadTree(long items) {
         return new QuadTreeSimple().init(items);
     }
